weeks/week-3/requirements.txt
azure-ai-projects
azure-ai-inference[prompts]
azure-ai-evaluation[remote]
azure-identity
azure-search-documents
azure-monitor-opentelemetry
pandas
python-dotenv
opentelemetry-api

